+++ src/test/createTestConfig.ts
@@ -51,6 +51,7 @@ function createTestConfig (projName: string, testFile: string, workspaceFolder?:
 	}
 
 	const launchArgs: string[] = []
+	launchArgs.push('--disable-gpu')
 	if (projName != 'DebugLines' &&
 		projName != 'proj3' &&
 		projName != 'proj4' &&
@@ -62,7 +63,7 @@ function createTestConfig (projName: string, testFile: string, workspaceFolder?:
 	}
 
 	if (insidersBuild) {
-		launchArgs.push('--enable-proposed-api=kherring.ablunit-test-provider')
+		launchArgs.push('--enable-proposed-api=kherring.ablunit-test-runner')
 	}
 
 	const retVal: ITestConfig = {